reorganize directory for  non-custmer components
