+++ src/components/customer/index.js
@@ -1,9 +1,7 @@
-export { default as BottomNavigationBar } from "./BottomBar";
 export { default as CheckIn } from "./CheckIn";
 export { default as OpenTabs } from "./OpenTabs";
 export { default as Tab } from "./Tab";
 export { default as TabItems } from "./TabItems";
-export { default as SplashScreen } from "./splashScreen";
 export { default as CustomerInfo } from "./CustomerInfo";
-export { default as Menu } from "./Menu";
-export { default as Logo } from "./Logo";
+export { default as Favorite } from "./Favorite";
+
