+++ src/components/main/Logo.js
