+++ src/components/main/index.js
@@ -1,2 +1,6 @@
 export { default as App } from "./App";
 export { default as Routes } from "./Routes";
+export { default as BottomNavigationBar } from "./BottomBar";
+export { default as SplashScreen } from "./splashScreen";
+export { default as Menu } from "./Menu";
+export { default as Logo } from "./Logo";
