+++ src/components/main/splashScreen.js
