+++ src/helpers/utilities.js
@@ -92,12 +92,12 @@ export function setErrorMsg(error) {
 
 export const isValidEmail = email => {
   return !!email.match(
-  /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+  /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
   )
 }
 
 export const isValidDOB = dob => {
   return !!dob.match(
-  /^(?:0[1-9]|1[12])[\/](?:0[1-9]|[12]\d|3[01])([\/])(?:19|20)\d\d$/
+  /^(?:0[1-9]|1[12])[/](?:0[1-9]|[12]\d|3[01])([/])(?:19|20)\d\d$/
   )
 }
